chore(zero): CI file size alerts (#3498)
This now sets up alerts when the file size grows by 5%

It also reports the file size as a PR comment so we can now ahead of merging to main about potential file size regressions.

This splits the YML into two files because PR needs slightly different flags. It is not optimal but having branching in yml is painful.