fix(zql): whereExists before where (#3552)
Fixes https://bugs.rocicorp.dev/issue/3417

If a query has a whereExists before a where, for example:

`z.query.issue.whereExists('creator').where('status', 'open');`

errors occur in the whereExists operator if a row is pushed to a source which does not match the where filter.

The issue arose because exists would receive a row `x` via push, it would then fetch with `startAt: x` to get the exists relationship for this row, but fetch would not contain row `x` as expected causing an assertion to fail.  This is because sources were applying filters to fetch but not push, so while `x` did not match the filters it would be pushed, but then excluded from fetch.

Updates MemorySource and TableSource to apply filters consistently on fetch and push. 

Renames `optionalFilters` to `filters` and, `appliedFilters` to `fullyAppliedFilters`.  If `fullyAppliedFilters` is `true`, filtering is fully handled by the sources, and the filters for where are not added to the pipeline.  

Also updates packages/zql/src/query/query-impl.query.test.ts to run against both MemorySource and TableSource (as it should have been).