fix(zql): `null` should not be allowed in `where` except with `IS`.
This originally worked but got broken with some restructuring in
query.ts. Adds some tests to make sure it doesn't break again.

Fixes: https://bugs.rocicorp.dev/issue/3426i

BREAKING CHANGE: This narrows the allowed types for where both
in all cases. For IS/IS NOT, null is only allowed if the column
type is optional. For other operators, null is now never allowed.