fix(zero): Add missing type exports (#3578)
https://bugs.rocicorp.dev/issue/3459

When a library imports `@rocicorp/zero` and uses `definePermission`,
and the library has a `tsconfig.json` that includes `declaration: true`,
then tsc complains that it cannot infer the type:

```
The inferred type of 'permissions' cannot be named without a reference
to '../node_modules/@rocicorp/zero/out/zero-protocol/src/ast'. This is
likely not portable. A type annotation is necessary.
```

The reason is that `definePermission` has a dependency on the
`Condition` type from `ast.ts` from `zero-protocol`, but that file is
never loaded by tsc so it does not know about this type.

The solution is to export a type dependency from `ast.ts`. That way
`tsc` is happy.

However, to be less magic we now export all the **transitive**
dependencies of `definePermission`. This leads to us eventually
exporting `AST` etc from `ast.ts`. This is not ideal. I didn't want
us to have to commit to the shape of the AST but we are already making
it accessible to the outside world via `definePermission` (even though
it was not very straight forward to get to it before).