fix(zero-react): Collapse onData with empty result
This ensures that if we get empty data we return a singleton. This
prevents the `useSyncExternalStore` from being triggered when the
`onData` goes from the default `[[], 'unknown']` to `[[], 'unknown']`.

This also deals with all 4 cases of the empty result to prevent extra
rerenders in react.

Also use WeakMap for test only exposure of allViews.

https://bugs.rocicorp.dev/issue/3456