feat(zero-react): remove extra empty data render
Every time we mount a component we do a render with no data then a render with data.

I made a change to fix this previously and it caused the zbugs issue list to render not data when going back.

This was not a bug in the change but rather a bug in our `use-element-size` hook which is now fixed.