refactor(zql): callable relationships (#3629)
The `Exist` operators `#fetchNodeForRow` is quite expensive and is a work around for the fact that `Node`'s relationships can only be consumed a single time.  In order to read the relevant relationship, `Exist` has to fetch from source the parent `Node`.   This fetch is pure overhead, we are fetching a row we already have.

This change updates `Node`'s relationships to be callable, so that they can be consumed multiple times.

`Node`'s type changes from 

```
export type Node = {
  row: Row;
  relationships: Record<string, Stream<Node>>;
};
```

to 

```
export type Node = {
  row: Row;
  relationships: Record<string, () => Stream<Node>>;
};
```

This change also updates `ChildChange` to have a `node: Node` rather than a `row: Row` so that the relationships needed by `Exist` are also available on `ChildChange`.  Now all `Change` types use `Node` rather than `Row`.

Finally this addresses an issue with cleanup where Node's filtered by `Exists` or `Filter` were not being drained (as they never reach view-apply-change which does the draining).  Added draining to cleanup in `Exists` and `Filter` for `Node`s not yielded to downstream.