perf(zero-solid): use reconcile (#3682)
Setting solid store state many times is very slow, even if it is within a Solid batch call.  In the hello-zero-solid app, if there are 3000 rows, processing the pushes during initial load takes 30 seconds. Nearly all of the time is in an internal solid function called unwrap.  
![image](https://github.com/user-attachments/assets/6ea3b512-bcc9-41f2-b8e3-4b3a1b57c7d1)
![image](https://github.com/user-attachments/assets/eaf76e61-dd0d-46e5-8fe4-2e5d14196de2)


The previous approach to solving this performance issue, https://github.com/rocicorp/mono/pull/3608, is incorrect cause we cannot defer processing pushes, pushes have to be processed as received, because the relationship streams will be incorrect if processing is deferred.  This was surfaced by issue https://bugs.rocicorp.dev/issue/3488
For comparison here is the performance of that approach:
<img width="1506" alt="image" src="https://github.com/user-attachments/assets/c82b5051-7176-4e90-96e1-e97b069ae63a" />

This new approach maintains a copy of the view as a plain JS object which is updated on each push as the push is received, and then uses solid's `reconcile` to update the store at the end of the transaction.
<img width="1425" alt="image" src="https://github.com/user-attachments/assets/ec56443b-12a1-4ed8-8164-b99f447988da" />
The downside is we are ~2xing memory (actually this depends on how much of the data is interned string, its probably actually much less than 2x), and we are diffing the views (which doesnt seem to match the solid way of precise reactivity, we are translating a stream of precise updates into a big diff).

The upside is its 
1. correct
2. event faster than the previous optimization 120 ms vs 970 ms.  Probably because in this test case the diff is trivial.