perf,fix(zero-solid): back to produce but pre-reading relationships and optimizing initial view build (#3701)
The `reconcile` optimization approach (https://github.com/rocicorp/mono/pull/3682) was incorrect (I misunderstood how reconcile works), as was revealed by the tests added here: https://github.com/rocicorp/mono/pull/3692.  

This approach goes back to using `produce`, but addresses the correctness issue by using an idea @tantaman came up with of reading the relationships of the changes at the time the push is received, but still queueing them to be applied to the solid store state on transaction completion.   We are careful to only read relationships in the changes actually used by `applyChanges`.

This solved the correctness problem, but while much faster than the 30 seconds prior to us trying to optimize this, the 3000 row initial load case in hello-zero-solid was still spending ~742ms in  SolidView#applyChanges.  
<img width="1434" alt="Pasted Graphic 2" src="https://github.com/user-attachments/assets/1179b9c4-90eb-4e3d-b0c7-e027ec51fb34" />


This is because the proxy used by solid `produce` is fairly slow.   To address this, I added an optimization for the case where the store's view is currently empty.  If the store is currently empty build up the view on a new plain old JS object root, which is returned for the new state.  This avoids building up large views from scratch using solid `produce`.  This brings the time for the 3000 row initial load case in hello-zero-solid to ~133ms.
<img width="1441" alt="Pasted Graphic 1" src="https://github.com/user-attachments/assets/0297b057-c826-420c-bdab-66ff2ee1f06d" />