perf(zero-solid): avoid temp arrays when building from empty (#3702)
This further optimizes the case of building from empty by avoiding allocating temp arrays.

Now we only have to allocate temp arrays for storing relationships between push and transaction commit when updating a non empty view, which is often smaller updates.

This didn't have an appreciable impact on the test case we've been using of 3000 rows (with 2, 1 child, relationships per row) in hello-zero-solid, but those rows all have very small non-nested relationships, so thats understandable.  I still think this is a worthwhile optimization despite it adding a bit of complexity.
<img width="1443" alt="image" src="https://github.com/user-attachments/assets/e7c13c3a-6a39-444d-aaad-0eab31d6b78f" />