feat(zero-pg): move `mutator` interface to `zql`
custom mutator interfaces are required on the client and server. They also depend on the `Query` type, hence the reason to move them to the `zql` package.

Note to self: zql includes `ivm` which isn't needed on the user's API server. The ZQL query builder interfaces are needed there, however. ZQL package could be split into `zql-interface` and `zql`.