chore(zero): Remove syncing of connected clients (#3792)
We used to sync the clientIDs of the connected clients. The idea was
that we would use this information to determine which clients we cannot
delete. However, this information is not used anywhere in the codebase
and is not necessary. This commit removes the syncing of the clientIDs.

This means that we can remove the clientPatches that we send.

Since we are not sending client patches we can also remove this state
from the CRV... however, since there might be multiple view syncers running we continue to write this state to the db. We have removed all the reads in the code now so we can remove the writes in the future when all the readers have migrated.