feat(zero-cache): Implement server side of deleteClients (#3813)
When the server receives a `DeleteClients` message, it will delete the client (and client group) from the CVR database. This will send `'desiredQueriesPatches'` patches to connected
clients that are affected by the deletions.

The server also confirms the deletion by sending a `'deleteClients'` back to the connected clients that requested the deletion.

Towards https://bugs.rocicorp.dev/issue/3476