fix(zql): Exist operator: remove cached size reuse across rows during push  (#3887)
The overlay solution to split push results in relationship's being inconsistent between Nodes at the same level of the tree during push. This means its not correct to reuse cached sizes across rows during push.

With this change initial hydration is really the only time that the cross row cache optimization is having an impact.