feat(zero-cache): Query TTL and eviction (#3850)
A query has a `ttl` and an `inactivatedAt` timestamp. This is a per
client state and stored in the `clientState` property of the
`ClientQueryRecord`. If a query is desired then it has an
`inactivatedAt` set to `undefined`. If a query is not desired then it
has an `inactivatedAt` set to a timestamp.

The actual removal of the query and the rows related to the query is
done in `#syncQueryPipelineSet`.

The scheduling of the removal based on the `ttl` + `inactivatedAt` is
done in `#scheduleExpireEviction` and it is done using a timer which is
set to the next expiration time.

The eviction of the inactive queries is done in `#evictInactiveQueries`.
If the row count is above the desired max row count we find the oldest
inactive query and remove it. This is done in a loop until the row count
is below the desired max row count or until there are no more inactive
queries.