fix(zero-client)!: Fix array view with multiple entries (#3922)
BREAKING CHANGE!

The code assumed that one could never get two adds of the same row.
However, this can happen when we have many to many relationships.

To deal with this we introduce a ref count to the array view. This
allows us to keep track of how many times a row has been added to the
view. When the ref count goes to zero we remove the row from the view.

This is a breaking change because `applyChange` now has a new required
parameter `refCountMap`. It will require updates to vue-zero and
vue-svelte.

https://bugs.rocicorp.dev/issue/3563