feat(custom-mutators): do not do IVM forking if custom mutators are not provided
If the user of Zero has not defined any custom mutators, do not fork the IVM sources
and return `undefined` instead.

This is mainly a safety thing. We want there to be a minimal chance of breaking
existing users while we're still stabilizing custom mutators. CRUD mutators
also do not need forks of the IVM sources since they do not issue reads during
mutations.