refactor(zql): move splitting of edits up to source (#3936)
This slightly simplifies take and join logic, but more importantly it avoids having to add fairly complex overlay logic in take and join to correctly handle split pushes. Instead the overlay logic for edits split into a remove and add is centralized at the source layer.

This is done by gathering the join keys and partition keys in builder and passing them to the source connection as a set called splitEditKeys. If the an edit modifies the values of any of these keys, then the source will push it as a remove followed by an add instead of as an edit.

Added asserts in join and take ensuring no unexpected edits are received.