fix(zero-client): Send updated TTL when changed  (#3972)
When we use `useQuery` (both Solid and React) we check if we already have a view for the query. If we do we reuse the view but update the ttl as needed.

To do this we have to add updateTTL to the AdvancedQuery. This in turn calls `updateServerQuery` on the `QueryDelegate` which in turn calls `update` on the `QueryManager`.