fix(zql): Fix ref count for edit in applyChange (#3977)
The edit branch was incomplete.


### Singular

We replace the old object with the new object transferring the ref count.


### Not Singular

If we edit the value of a non primary column, we transfer the ref count to the new object.

If we edit the value of a primary column, we need to decrement the ref count of the old object and increment the ref count of the new object.