chore(z2s): type conversions for query results (#4110)
This make z2s output TEXT containing json in order to minimize the chance of any type converters installed on the customers pg client library incorreclty changing the results.

It also correctly converts `timestamp`, `timestamptz` and `date` columns to `number` in the json output.  To be able to do this we need to add more information to the zero schema.  Currently these column types are all represented simply as `type: 'number'` in zero schema.  To determine that these numbers are actually timestamps/dates we'd have to fetch the upstream db's schema, and since the api server may be a stateless lambda this might need to be done effectively on every push.  Instead we disambiguate these types in the zero schema.

'bigint' handling to be done in a follow up.

This change has backwards compatibility, it will still work with schema's that use 'number' for timestamps and dates, these schema's just won't work with z2s.