feat(zero)!: Add options to Query run (#4226)
BREAKING CHANGE!

```ts
type RunOptions = {
  type: 'complete' | 'unknown';
}

interface Query<...> {
  ...
  run(options?: RunOptions): Promise<...>;
}
```

Defaults to `complete` for all queries except for the query used inside
custom mutations which defaults to `unknown`.

This is breaking because the `run` method of the `Query` class now waits
for `complete` results by default. This can lead to the promise returned
never resolving if the `server` is not set (or we are offline).

Another gotcha is that if you call run with `unknown` you get both
unknown and complete results but there is no way to tell which is
which.

https://bugs.rocicorp.dev/issue/3243