chore(zero-client)!: Change run default to unknown (#4234)
BREAKING CHANGE -- This is a breaking change from a previous change but
it restores the old 0.18 behavior where run returns the snapshot by default.

This simplifies the code since run always defaults to unknown now.

We still throw if you try to pass complete to a run inside a custom
mutator.