perf(zql): optimization for sparce/empty OR branches (#4339)
**Problem**
Fetching a pipeline with an OR is very inefficient if one or more branches are sparse or empty.  

This is because of how `FanIn` merges fetch `Stream`s.  It essentially uses merge-sort, asking each `Stream` for its next `Node `to determine the next `Node` in sort order.  A sparse or empty branch may process every (or nearly every) row in its source table to answer that its `Stream` is empty, even if another branch could satisfy the limit of the query processing far fewer rows.

For example lets say the condition of a query is `or(cmp('foo', 'bar'), and(cmp('fuzzy', 'wuzzy'), exists('baz')))`, where `cmp('foo', 'bar')` is always false but `and(cmp('fuzzy', 'wuzzy'), exists('baz'))` has many matches early in the sort order.  We need a complex query with exists, cause otherwise we have an optimization that will hoist the whole filter to the SqlLite query. We will fetch from SqlLite `or(cmp('foo', 'bar'), cmp('fuzzy', 'wuzzy'))`, which will effectively return  `cmp('fuzzy', 'wuzzy')`, since `cmp('foo', 'bar')` is always false. FanIn/FanOut will do this fetch for each branch of the or.  In order to merge in the right sort order it will ask for the next item from branch `cmp('foo', 'bar')`, and the next item from branch `and(cmp('fuzzy', 'wuzzy'), exists('baz'))`, to compare them. To get the next item from branch `cmp('fuzzy', 'wuzzy')`, we fetch `or(cmp('foo', 'bar'), cmp('fuzzy', 'wuzzy'))` from SqlLite which has many rows, but then the Filter operator for branch `cmp('fuzzy', 'wuzzy')` discards all the rows.  We effectively process all of `cmp('fuzzy', 'wuzzy')` (without respect to limit, so worst case the entire table) to answer that the next item for `cmp('foo', 'bar')` doesn't exist. 

There are two significant inefficiencies here:
1. a query to SqlLite is made per branch of OR.  Each query has to be broad enough to return all rows which may match any branch of the OR.
2. to correctly order/dedup the merged output of these branches, many more rows than necessary (in many case all rows of a table) are processed if one or more branch is empty/sparse.

**Solution**

Eliminate doing a query per branch.  Do a single query that is broad enough to return all rows which may match any branch of the OR.

Process the nodes returned from this single query in order, filtering them one at a time, eliminating the need to merge at all.

This is accomplished by introducing the new abstraction of a `FilterOperator`.  The `where` clause of a ZQL query is implemented using a sub-graph of `FilterOperators`.  This sub-graph starts with a `FilterStart` operator, that adapts from the normal `Operator` `Output`, to the `FilterOperator` `FilterInput`, and ends with a `FilterEnd` operator that adapts from a `FilterOperator` `FilterOutput` to a normal `Operator` `Input`.  `FilterOperator'`s do not have `fetch` or `cleanup` instead they have a `filter(node: Node, cleanup: boolean): boolean` method.  They also have `push` which is just like normal `Operator` push.  There are currently 4 `FilterOperator`s: `Filter`, `Exist`, `FanIn` and `FanOut`.    Not having a `fetch` means these `FilterOperator`'s cannot modify Node rows or relationships, but they shouldn't, they should just filter  This new abstraction allows much more efficient processing of `fetch` for an OR condition.  
_Fetch with a FilterOperator subgraph_
When a `FilterEnd` receives a fetch, it fetches from its corresponding `FilterStart` at the start of the `FilterOperator` subgraph, which fetches from it's `Input` (which is a normal `Operator`), it then asks its `FilterOutput` to filter each node, yielding to `FilterEnd` those Node's which pass the filter. 
_Push with a FilterOperator subgraph_
Push works as it does today.  `FilterStart` receives a push, which pushes to its `FilterOutput` which pushes to its `FilterOutput`(s) (maybe multiple, as `FanOut` can branch out pushes), until the push conditionally (it may be filtered out) reaches `FilterEnd`, which pushes it to it normal `Operator` `Output`.

Note this code structure can also support correctly and efficiently processing nested ORs (i.e. nested `FanIn` / `FanOut`), making it possible to not convert `where` conditions to DNF.


Perf impact on a 'OR with empty branch and limit with exists' scenario  using the chinook tracks data set yielded a 100x speed up.
```
 ✓  zql-benchmarks/pg-17  src/chinook-hydration.bench.ts > OR with empty branch and limit with exists 1816ms
     name           hz     min      max    mean     p75      p99     p995     p999     rme  samples
   · zql     29,029.27  0.0238   3.0766  0.0344  0.0268   0.0718   0.1245   2.1104  ±5.83%    14515  [35.78x] ⇑
     zql        811.28  1.0302   2.7010  1.2326  1.1540   2.5545   2.6407   2.7010  ±3.09%      406  (baseline)
   · zqlite  10,963.91  0.0690   2.9097  0.0912  0.0767   0.2030   1.7859   2.3478  ±4.99%     5482  [99.17x] ⇑
     zqlite     110.56  8.3685  10.5341  9.0450  9.7990  10.5341  10.5341  10.5341  ±1.80%       56  (baseline)
   · zpg        898.83  0.7937   4.8358  1.1126  1.1835   2.0152   2.6168   4.8358  ±2.44%      450  [1.21x] ⇑
     zpg        744.30  0.8805   6.7553  1.3435  1.3058   4.4409   5.4718   6.7553  ±5.07%      373  (baseline)
```